Move Logic to Stopwatch component
